Defaulted ResponsiblePeople to empty list on load
This stems from a similar problem to Trading Premises, where the user could potentially submit an application with no Trading Premises or Responsible People.

We've already put the fix in for Trading Premises, this one is for Responsible People.

Also added some tests, which were missing from the first implementation.