New preAppComplete flag on BusinessMatching model
This flag was added as there's currently no other way to tell whether the user has initially completed the business matching section or not.
This was playing havoc with the 'return to application progress' links inside business matching, which was previously based on the 'hasAccepted'
flag being true. Problem is, the state of that flag changes as you go through and edit the questions in business matching, meaning that the
registration progress link would appear to show randomly on some pages and not others. This should hopefully fix that.

I've taken care to make sure that the flag is defaulted to true when the data is read from API 5.