AMLS-4079 Changed data migration strategy
Previously, data was migrated when loaded and the data was primarily loaded from the mongo cache. This commit changes that so that the data is primarily loaded from S4L, and the data is migrated slowly on save.

At some point later (after 28 days has passed), we're going to have to update this so that the mongo store becomes the primary cache and the data is dropped from S4L.