HOTFIX: Remove cache from the right place on redirect
Since migrating to Mongo, it does not make sense to the use the ShortLivedCache.remove method, as this only removes data from Save4Later.

This commit changes that to remove it from the data cache connector, which will ultimately remove the data from Mongo and Save4Later.

This could potentially fix a live issue we're seeing where the start page constantly redirects to itself.